Problem: zproc_kill may kill not only this process but other child processes and parent process too
```int kill( pid_t pid, int sig )```
```If pid is equal to zero, kill() sends the signal sig to all processes whose process group ID is equal to that of the sender, except for those to which the sender does not have the appropriate authority to send a signal```

Solution: if self->pid is still not known -> do nothing

+++ src/zproc.c
@@ -522,7 +522,7 @@ s_zproc_execve (zproc_t *self)
     char *filename = (char*) zlist_first (self->args);
     self->pid = fork ();
     if (self->pid == 0) {
-
+        // Child
         if (self->stdinpipe [0] != 0) {
             int o_flags = fcntl (self->stdinpipe [0], F_GETFL);
             int n_flags = o_flags & (~O_NONBLOCK);
@@ -585,10 +585,12 @@ s_zproc_execve (zproc_t *self)
     }
     else
     if (self->pid == -1) {
+        // Error (still in parent)
         zsys_error ("error fork: %s", strerror (errno));
         exit (EXIT_FAILURE);
     }
     else {
+        // Parent
         if (self->verbose)
             zsys_debug ("process %s with pid %d started", filename, self->pid);
 
@@ -600,12 +602,12 @@ s_zproc_execve (zproc_t *self)
         // add a handler for read end of stdout
         if (self->stdoutpipe [1] != -1) {
             s_zproc_addfd (self, self->stdoutpipe [0], zpair_write (self->stdoutpair), ZMQ_POLLIN);
-            close(self->stdoutpipe[1]);
+            close (self->stdoutpipe[1]);
         }
         // add a handler for read end of stderr
         if (self->stderrpipe [1] != -1) {
             s_zproc_addfd (self, self->stderrpipe [0], zpair_write (self->stderrpair), ZMQ_POLLIN);
-            close(self->stderrpipe[1]);
+            close (self->stderrpipe[1]);
         }
     }
 
@@ -768,10 +770,12 @@ zproc_kill (zproc_t *self, int signum) {
     zsys_debug ("zproc_kill not implemented on Windows");
     return;
 #else
-    int r = kill (self->pid, signum);
-    if (r != 0)
-        zsys_error ("kill of pid=%d failed: %s", self->pid, strerror (errno));
-    zproc_wait (self, false);
+    if (zproc_pid (self) > 0) {
+        int r = kill (self->pid, signum);
+        if (r != 0)
+            zsys_error ("kill of pid=%d failed: %s", self->pid, strerror (errno));
+        zproc_wait (self, false);
+    }
 #endif
 }
